refactor(database.py): remove database module
Remove module and migrate its functionality to utils
