+++ invenio_rdm_pure/source/utils.py
@@ -14,14 +14,39 @@ from pathlib import Path
 from typing import List
 
 from flask import current_app
+from flask_security.utils import hash_password
+from invenio_db import db
 
-from ..setup import (
-    data_files_name,
-    email_message,
-    email_smtp_port,
-    email_smtp_server,
-    pure_uuid_length,
+
+def get_user_id(user_email: str, user_password: str):
+    """Get the userId of the user.
+
+    In case the user doesn't exist yet,
+    create it with given credentials.
+    """
+    datastore = current_app.extensions["security"].datastore
+    if datastore is not None:
+        user = datastore.get_user(user_email)
+        if not user:
+            user = datastore.create_user(
+                email=user_email,
+                password=hash_password(user_password),
+                active=True,
 )
+            db.session.commit()
+        return user.id
+
+
+def make_user_admin(self, id_or_email: str) -> None:
+    """Gives the user with given id or email administrator rights."""
+    return None  # FIXME: Method stub'd until auxiliary methods are implemented.
+    datastore = current_app.extensions["security"].datastore
+    if datastore is not None:
+        invenio_pure_user = datastore.get_user(
+            id_or_email
+        )  # FIXME: Not implemented yet.
+        admin_role = datastore.find_role("admin")  # FIXME: Not implemented yet.
+        datastore.add_role_to_user(invenio_pure_user, admin_role)
 
 
 def load_file_as_string(path):
